Add Kotlin Client Generation (#24)
* add .idea files to git

* messing around

* asdf

* asdf

* progress for sse support

* get retry and event ID stuff working

* add some initial tests and model generation

* start setting up kotlin rpc generation

* start working on service def

* further kotlin progress

* update libs

* nx migrate

* updates

* finish initial implementation

* start setting up kotlin tests

* fix cwd err

* update libs

* setup test client

* add first integration-test for kotlin

* update bun types

* start messing with standard tests and stream tests

* in progress kotlin fixes for discriminated unions

* pnpm i

* experimenting with kotlin polymorphic serialization

* fixing errs

* revert to old test structure

* asdf

* adfas

* change path

* up to date

* fix type issues

* further kotlin progress

* get recursive object stuff

* kotlin recursive stuff

* further kotlin progress

* improve union generation

* rebuild

* kotlin generation progress

* rebuild client

* update build

* updates

* asdf

* fix bug in sealed class generation

* cleaning up kotlin codegen

* reference string utils from kotlinxserialization

* todo create more standardized tests

* add generic serialization testing resources

* improve generated test files

* progress

* more standardized to and from json methods

* add tests for mismatch json types

* check gradlew and gradlew.bat into git

* start mapping out recursive object parsing in kotlin

* kotlin client progress

* safdadf

* add missing tests

* starting over on kotlin codegen

* further kotlin progress

* finish http services

* fix output errors

* fix sse cancellation issues

* improving kotlin tests

* fix sse implementation issues

* remove WS notes

* get sse working with large payloads

+++ tests/server/src/procedures/tests/streamConnectionErrorTest.rpc.ts
@@ -5,7 +5,7 @@ export default defineEventStreamRpc({
     description:
         "This route will always return an error. The client should automatically retry with exponential backoff.",
     params: a.object("StreamConnectionErrorTestParams", {
-        statusCode: a.uint16(),
+        statusCode: a.int32(),
         statusMessage: a.string(),
     }),
     response: a.object("StreamConnectionErrorTestResponse", {
